pyproject.toml
[project]
name = "rickle"
version = "1.2.4"
description = "Python tools for working with configuration files"
authors = [
    { name = "Zipfian.sh", email = "about@zipfian.sh" }
]
license = { text = "Apache 2.0" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml", "requests", "tomli;python_version<'3.11'", "tomli-w"
]
keywords = ['Pickle', 'Python', 'config', 'configuration', 'YAML', 'JSON', 'TOML', 'XML', 'INI', 'ENV']
classifiers = [
    "Development Status :: 5 - Production/Stable",
    'Intended Audience :: Developers',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
    'Topic :: Software Development',
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
net = [
    'twisted',
    'pyopenssl'
]
xml = ['xmltodict']
dotenv = ['python-dotenv']
validators = ["py.validator"]
jsonschema = ["jsonschema[format]"]
full = [
    "py.validator",
    'python-dotenv',
    'xmltodict',
    'twisted',
    'pyopenssl',
    "jsonschema[format]"
]

[project.scripts]
rickle = "rickle.cli:main"

[project.urls]
Homepage = "https://zipfian.sh/projects/rickle.html"
Documentation = "https://zipfian.sh/docs/rickle/"
Repository = "https://github.com/zipfian-sh/rickle"
Issues = "https://github.com/zipfian-sh/rickle/issues"
Changelog = "https://zipfian.sh/docs/rickle/changelog/"

[tool.poetry.scripts]
unittest = 'build_utils:all_unit_tests'
bumpver = 'build_utils:bump_version_patch'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
jsonschema = "^4.23.0"
coverage = "^7.6.10"
pyopenssl = "^25.0.0"
twisted = "^24.11.0"
python-dotenv = "^1.0.1"
xmltodict = "^0.14.2"
py-validator = "^0.21.1"
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.14"
mkdocstrings = {extras = ["python"], version = "^0.29.1"}

